[ALPS05066979] ap: Fix coverity
CERT-C Error Handling:
 - P2p wext command snprintf * 3
CERT-C Characters and Strings:
 - P2P State * 4, p2p index * 6, ch rank list * 2
Dereference before null check:
 - Channel notify
Uninitialized scalar variable:
 - Init channel list * 2

MTK-Commit-Id: a3acd2292369f73d7770d034af7ab08df125d71c

CR-Id: ALPS05066979
Feature: [Module]Wi-Fi Driver

+++ os/linux/gl_p2p_init.c
@@ -217,7 +217,7 @@ void p2pSetMode(IN uint8_t ucAPMode)
 /*---------------------------------------------------------------------------*/
 u_int8_t p2pRemove(struct GLUE_INFO *prGlueInfo)
 {
-	int idx = 0;
+	u_int8_t idx = 0;
 
 	if (prGlueInfo->prAdapter->fgIsP2PRegistered == FALSE) {
 		DBGLOG(P2P, INFO, "p2p is not registered\n");
